requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.44.1
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
Eel==0.17.0
evdev==1.7.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
idna==3.7
iso8601==2.1.0
lml==0.1.0
lxml==5.2.2
packaging==24.1
pefile==2023.2.7
pyasn1==0.6.0
pycparser==2.22
pyexcel-ezodf==0.3.4
pyexcel-io==0.6.6
pyexcel-ods3==0.6.1
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pynput==1.7.7
pyparsing==3.1.2
pyserial==3.5
python-xlib==0.33
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
setuptools==70.2.0
six==1.16.0
urllib3==2.2.2
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.4.post2
